frontend: Refactor code into components for easier maintenance
