+++ frontend/src/App.jsx
@@ -4,19 +4,11 @@ import {
   Routes,
   Route,
   Link,
-  useNavigate,
   useLocation,
 } from "react-router-dom";
-import { Line } from "react-chartjs-2";
-// DO NOT REMOVE
-// necessary to avoid "category is not a registered scale" error.
-import { Chart as ChartJS } from "chart.js/auto";
-import { Chart } from "react-chartjs-2";
-import { format } from "date-fns";
-import { AgGridReact } from "ag-grid-react"; // React Grid Logic
-import "ag-grid-community/styles/ag-grid.css"; // Core CSS
-import "ag-grid-community/styles/ag-theme-quartz.css"; // Theme
 import "./App.css";
+import { Login, LogOut, LoginButton } from "./components/Login.jsx";
+import { Dashboard, SingleResult } from "./components/Dashboard.jsx";
 
 const LoggedInContext = createContext(false);
 
@@ -38,138 +30,6 @@ async function loginUser(credentials) {
     body: JSON.stringify(credentials),
   }).then((data) => data.json());
 }
-
-const Login = ({ loggedIn, setLoggedIn }) => {
-  const [username, setUsername] = useState();
-  const [password, setPassword] = useState();
-
-  const githubSubmit = async (e) => {
-    // e.preventDefault();
-    // console.log("Github submit");
-    // const data = await fetch("http://localhost/api/v0/auth/github/authorize")
-    //   .then((response) => response.json())
-    //   .then((url) => url["authorization_url"])
-    //   .then((url) => {
-    //     console.log(url);
-    //     window.location.href = url;
-    //     setLoggedIn(true);
-    //     localStorage.setItem("loggedIn", "true");
-    //   })
-    //   .catch((error) => console.log(error));
-  };
-
-  const navigate = useNavigate();
-  const authSubmit = async (e) => {
-    e.preventDefault();
-    console.log("Auth submit: " + username + " " + password);
-    let credentialsData = new URLSearchParams();
-    credentialsData.append("username", username);
-    credentialsData.append("password", password);
-
-    const data = await fetch("/api/v0/auth/jwt/login", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded",
-      },
-      body: credentialsData,
-    })
-      .then((response) => response.json())
-      .then((body) => {
-        setLoggedIn(true);
-        localStorage.setItem("loggedIn", "true");
-        localStorage.setItem("token", body["access_token"]);
-        try {
-          navigate("/");
-        } catch (error) {
-          console.log(error);
-        }
-      })
-      .catch((error) => console.log(error));
-  };
-
-  return (
-    <div className="row mt-5">
-      <div className="col-sm-3 offset-sm-4">
-        <form onSubmit={authSubmit}>
-          <div className="mb-3">
-            <label htmlFor="exampleInputEmail1" className="form-label">
-              Email address
-            </label>
-            <input
-              type="text"
-              className="form-control"
-              id="exampleInputEmail1"
-              onChange={(e) => setUsername(e.target.value)}
-            />
-            <div className="mb-3">
-              <label htmlFor="exampleInputPassword1" className="form-label">
-                Password
-              </label>
-              <input
-                type="password"
-                className="form-control"
-                id="exampleInputPassword1"
-                onChange={(e) => setPassword(e.target.value)}
-              />
-            </div>
-          </div>
-          <div className="text-center mt-2">
-            <button type="submit" className="btn btn-success">
-              Submit
-            </button>
-          </div>
-        </form>
-        {/* <div className="text-center mt-5">
-          <button className="btn btn-success" onClick={githubSubmit}>
-            Log in with GitHub
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              width="16"
-              height="16"
-              fill="currentColor"
-              className="bi bi-github"
-              viewBox="0 0 16 16"
-            >
-              <path d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27.68 0 1.36.09 2 .27 1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.012 8.012 0 0 0 16 8c0-4.42-3.58-8-8-8" />
-            </svg>
-          </button>
-        </div> */}
-      </div>
-    </div>
-  );
-};
-
-const LogOut = ({ setLoggedIn }) => {
-  const handleLogoutClick = () => {
-    console.log("Setting logged in to false");
-    setLoggedIn(false);
-    localStorage.setItem("loggedIn", "false");
-  };
-  return (
-    <>
-      <button className="btn btn-success" onClick={handleLogoutClick}>
-        Log Out
-      </button>
-    </>
-  );
-};
-const LoginButton = ({ loggedIn, setLoggedIn }) => {
-  return (
-    <>
-      <Link
-        to="/login"
-        className="btn btn-success"
-        loggedIn={loggedIn}
-        setLoggedIn={setLoggedIn}
-      >
-        Log In
-      </Link>
-      {/* <a href="/foobar" className="btn btn-success" type="submit">
-          Sign Up
-        </a> */}
-    </>
-  );
-};
 
 const NavigationItems = () => {
   return (
@@ -331,357 +191,6 @@ const Root = ({ loggedIn }) => {
           </>
         )}
       </div>
-    </>
-  );
-};
-
-const parseTimestamp = (t) => {
-  const utcSeconds = t;
-  var d = new Date(0);
-  d.setUTCSeconds(utcSeconds);
-  return format(d, "yyyy-MM-dd HH:mm");
-};
-
-const formatCommit = (commit, commit_msg) => {
-  // Limit the git commit sha to 12 characters to improve readability
-  return commit.substring(0, 12) + ' ("' + commit_msg + '")';
-};
-
-const ChangePointSummaryTable = ({ changeData }) => {
-  var rowData = [];
-
-  Object.entries(changeData).forEach(([testName, value]) => {
-    value.forEach((changePoint) => {
-      console.log(changePoint);
-      const changes = changePoint["changes"];
-      console.log(changes);
-      changes.map((change) => {
-        const commit = changePoint["attributes"]["git_commit"][0];
-        const commit_msg = changePoint["attributes"]["commit_msg"][0];
-        const repo = changePoint["attributes"]["git_repo"][0];
-        rowData.push({
-          date: parseTimestamp(changePoint["time"]),
-          commit: { commit, commit_msg, repo },
-          metric: change["metric"],
-          change: change["forward_change_percent"] + "%",
-        });
-      });
-    });
-  });
-
-  const colDefs = [
-    { field: "date" },
-    { field: "metric" },
-    { field: "change" },
-    {
-      field: "commit",
-      cellRenderer: (params) => {
-        const { commit, commit_msg, repo } = params.value;
-        const url = repo + "/commit/" + commit;
-        const text = formatCommit(commit, commit_msg);
-        return (
-          <a href={url} target="_blank">
-            {text}
-          </a>
-        );
-      },
-    },
-  ];
-
-  return (
-    <>
-      <div className="ag-theme-quartz" style={{ height: 500, width: 900 }}>
-        <AgGridReact rowData={rowData} columnDefs={colDefs} pagination={true} />
-      </div>
-    </>
-  );
-};
-
-const SingleResult = () => {
-  const location = useLocation();
-  const testName = location.state.testName;
-  const [loading, setLoading] = useState(false);
-  const [displayData, setDisplayData] = useState([]);
-  const [changePointData, setChangePointData] = useState([]);
-  const fetchData = async () => {
-    const results = await fetch("/api/v0/result/" + testName, {
-      headers: {
-        "Content-type": "application/json",
-        Authorization: "Bearer " + localStorage.getItem("token"),
-      },
-    });
-    const resultData = await results.json();
-    resultData.sort((a, b) => {
-      return a.timestamp - b.timestamp;
-    });
-    setDisplayData(resultData);
-
-    const changes = await fetch("/api/v0/result/" + testName + "/changes", {
-      headers: {
-        "Content-type": "application/json",
-        Authorization: "Bearer " + localStorage.getItem("token"),
-      },
-    });
-    const changeData = await changes.json();
-    setChangePointData(changeData);
-  };
-
-  const parseData = (data, metricName) => {
-    console.log(data);
-    const value_map = data.map(
-      (result) =>
-        result.metrics
-          .filter((metric) => metric.name === metricName)
-          .map((metric) => metric.value)[0]
-    );
-    console.log(value_map);
-    return value_map;
-  };
-
-  useEffect(() => {
-    setLoading(true);
-    fetchData().finally(() => {
-      setLoading(false);
-    });
-  }, []);
-
-  console.log("Display data");
-  console.log(displayData);
-
-  const timestamps = displayData.map((result) => {
-    return result.timestamp;
-  });
-
-  var metricMap = [];
-  displayData.map((result) => {
-    result.metrics.map((metric) => {
-      metricMap.push({ name: metric.name, unit: metric.unit });
-    });
-  });
-
-  // Only want unique metric names
-  var unique = metricMap.reduce((a, b) => {
-    if (a.findIndex((x) => x.name === b.name) === -1) {
-      return a.concat([b]);
-    } else {
-      return a;
-    }
-  }, []);
-  console.log("unique: " + unique);
-
-  // {'testName':
-  //    [{
-  //      'time': 123,
-  //      'changes': [{'forward_change_percent': 900, 'metric': 'metric1'}]
-  //    }]
-  // }
-  const changePointTimes = [];
-
-  // TODO(mfleming) Assumes a single testName but must handle multiple
-  // tests in the future.
-  Object.entries(changePointData).forEach(([testName, value]) => {
-    value.forEach((changePoint) => {
-      const metrics = changePoint["changes"].map((change) => {
-        return change["metric"];
-      });
-      console.log(metrics);
-      const t = changePoint["time"];
-      changePointTimes.push({ t, metrics });
-    });
-  });
-
-  var changePointIndexes = [];
-  timestamps.map((timestamp, index) => {
-    changePointTimes.map((change) => {
-      const { t, metrics } = change;
-      if (t !== timestamp) {
-        return;
-      }
-      changePointIndexes.push({ index, metrics });
-    });
-  });
-
-  const drawLineChart = (metric) => {
-    const metricName = metric["name"];
-    const metricUnit = metric["unit"];
-    return (
-      <>
-        <h6 className="text-center">
-          {testName}: {metricName}
-        </h6>
-        <Line
-          datasetIdKey="foo"
-          data={{
-            labels: timestamps.map(parseTimestamp),
-            datasets: [
-              {
-                id: 1,
-                label: metricName,
-                data: parseData(displayData, metricName),
-                pointRadius: (context) => {
-                  const c = changePointIndexes;
-                  const entry = changePointIndexes.find((element) => {
-                    return (
-                      element.metrics.includes(metricName) &&
-                      element.index === context.dataIndex
-                    );
-                  });
-                  return entry ? 8 : 0;
-                },
-              },
-            ],
-          }}
-          options={{
-            scales: {
-              x: {
-                grid: {
-                  display: false,
-                },
-              },
-            },
-            plugins: {
-              legend: {
-                display: false,
-              },
-              interaction: {
-                intersect: false,
-                mode: "index",
-              },
-              tooltip: {
-                displayColors: false,
-                callbacks: {
-                  label: (context) => {
-                    var labelArray = ["value: " + context.raw + metricUnit];
-
-                    // Search in changePointData for this timestamp and metric
-                    const timestamp = timestamps[context.dataIndex];
-                    Object.entries(changePointData).forEach(
-                      ([testName, value]) => {
-                        value.forEach((changePoint) => {
-                          if (changePoint["time"] === timestamp) {
-                            labelArray.push("");
-
-                            // Add all change point attributes to the label
-                            changePoint["changes"].forEach((change) => {
-                              if (change["metric"] !== metricName) {
-                                return;
-                              }
-
-                              Object.entries(change).map(([key, value]) => {
-                                if (key === "metric") {
-                                  return;
-                                }
-
-                                var label = key + ": " + value;
-                                if (key.includes("percent")) {
-                                  label = label + "%";
-                                }
-
-                                labelArray.push(label);
-                              });
-                            });
-                          }
-                        });
-                      }
-                    );
-
-                    return labelArray;
-                  },
-                },
-                intersect: false,
-              },
-            },
-          }}
-        />
-      </>
-    );
-  };
-
-  return (
-    <>
-      {loading ? (
-        <div>Loading</div>
-      ) : (
-        <>
-          <nav className="navbar navbar-expand-lg">
-            <div className="container-fluid">
-              <nav aria-label="breadcrumb">
-                <ol className="breadcrumb">
-                  <li className="breadcrumb-item">
-                    <Link to="/">Tests</Link>
-                  </li>
-                  <li className="breadcrumb-item active" aria-current="page">
-                    {testName}
-                  </li>
-                </ol>
-              </nav>
-            </div>
-          </nav>
-          <div className="container">
-            <div className="row justify-content-center">
-              <ChangePointSummaryTable changeData={changePointData} />
-            </div>
-            <div className="row">{unique.map(drawLineChart)}</div>
-          </div>
-        </>
-      )}
-    </>
-  );
-};
-
-const Dashboard = () => {
-  const [loading, setLoading] = useState(false);
-  const [testNames, setTestNames] = useState([]);
-
-  const fetchData = async () => {
-    const response = await fetch("/api/v0/results", {
-      headers: {
-        "Content-type": "application/json",
-        Authorization: "Bearer " + localStorage.getItem("token"),
-      },
-    });
-    const resultData = await response.json();
-    resultData.map((element) => {
-      const test_name = element.test_name;
-      console.log(test_name);
-      setTestNames((prevState) => [...prevState, test_name]);
-    });
-  };
-  useEffect(() => {
-    setLoading(true);
-    fetchData().finally(() => {
-      setLoading(false);
-    });
-  }, []);
-
-  return (
-    <>
-      {loading ? (
-        <div>Loading</div>
-      ) : (
-        <div className="container">
-          <div className="card">
-            <div className="card-header">Please select a test</div>
-            <div className="card-body">
-              <ul className="list-group list-group-flush">
-                {testNames.map((name, i) => {
-                  return (
-                    <li className="list-group-item">
-                      <Link
-                        to="/result"
-                        testName={name}
-                        state={{ testName: name }}
-                      >
-                        {name}
-                      </Link>
-                    </li>
-                  );
-                })}
-              </ul>
-            </div>
-          </div>
-        </div>
-      )}
     </>
   );
 };
